~Renamed setup.py to initialization.py. ~Added requirements.txt. ~Edited README.md.
