+++ requirements.txt
@@ -1,0 +1,7 @@
+altgraph==0.17
+future==0.18.2
+pefile==2019.4.18
+pyinstaller==4.1
+pyinstaller-hooks-contrib==2020.10
+pywin32-ctypes==0.2.0
+tqdm==4.54.1
